Implement Gershwin word definitions as Clojure functions
Current impl. highly simplistic, parses word definition (symbol for name, stack effect, optional docstring added to metadata, and a definition body), creates a Clojure `def` form with the word definition body wrapped as a function. Every form in the word definition that is not a Gershwin word is simply conj'ed onto the data stack; every form that is a word is wrapped in a call to invoke it.