Support Gershwin words that are not name-munged
This, along with with-gershwin, will allow properly handling functions that need to be definedand attached to vars via purely Clojure features (e.g., protocols). Before this commit, Gershwin extensions to the compiler only looked for name-munged vars suffixed with __GWN__ when trying to resolve a symbol as a word. This change falls back to checking the non-name-munged var for {:word true} metadata as well.